[terra-functional-testing] Remove ENVs to use CLI options (#583)
* Add additional CLI options.

* Remove usage of env in favor of CLI options.

* Use updateScreenshot that is passed down in the VR service options.

* Pass down services to Launcher.

* Remove baseUrl, hostname, and port CLI options.

* Add back the empty line that was removed.

* Enforce utilizing test runner with cli options.

* Disable selenium service if SELENIUM_GRID_URL is provided.  Remove "strict" api to allow any any CLI option.

* Check for ENVs inside the test runner.

* Remove the brackets if BROWSERS is set as an array like ['chrome, firefox, ie'] or ['chrome','firefox','ie'] in Jenkinsfile.

* Update changelog.

* site should be returned as a string instead of array.

+++ packages/terra-functional-testing/src/services/wdio-selenium-docker-service.js
@@ -10,28 +10,33 @@ const logger = new Logger({ prefix: '[terra-functional-testing:wdio-selenium-doc
 const exec = util.promisify(childProcess.exec);
 
 class SeleniumDockerService {
-  constructor(options = {}) {
-    const { version } = options;
+  /**
+   * Service constructor.
+   * @param {Object} _options - The options specific to this service.
+   * @param {Object} _capabilities - The list of capabilities details.
+   * @param {Object} config - The object containing the wdio configuration and options defined in the terra-cli test runner.
+   */
+  constructor(_options, _capabilities, config = {}) {
+    const { launcherOptions } = config;
+    const { disableSeleniumService, keepAliveSeleniumDockerService, version } = launcherOptions || {};
 
+    this.disableSeleniumService = disableSeleniumService === true;
+    this.keepAliveSeleniumDockerService = keepAliveSeleniumDockerService === true;
     this.version = version;
   }
 
   /**
    * Gets executed once before all workers get launched.
-   * @param {Object} config - The wdio configuration object.
    */
-  async onPrepare(config) {
-    const { launcherOptions } = config;
-    const { keepAliveSeleniumDockerService } = launcherOptions || {};
-
-    this.keepAliveSeleniumDockerService = keepAliveSeleniumDockerService === true;
-
-    try {
-      // Verify docker is installed.
-      await exec('docker -v');
-      await this.startSeleniumHub();
-    } catch (error) {
-      throw new SevereServiceError(error);
+  async onPrepare() {
+    if (!this.disableSeleniumService) {
+      try {
+        // Verify docker is installed.
+        await exec('docker -v');
+        await this.startSeleniumHub();
+      } catch (error) {
+        throw new SevereServiceError(error);
+      }
     }
   }
 
@@ -117,7 +122,7 @@ class SeleniumDockerService {
     // When multiple test sessions are executed sequentially as specified in the WDIO script in the package.json file, the keepAliveSeleniumDockerService cli option
     // is used to indicate not to remove the currently deployed docker stack upon test completion as it will be used again for the next test session.
     // The docker stack is expected to be removed by the last test session when no keepAliveSeleniumDockerService cli option is specified.
-    if (!this.keepAliveSeleniumDockerService) {
+    if (!this.keepAliveSeleniumDockerService && !this.disableSeleniumService) {
       logger.info('Shutting down the docker selenium hub...');
 
       await exec(`docker-compose -f ${this.getDockerComposeFilePath()} down`);
